Beginnings of jump mode
This kind of works for at least translation, and maybe will work for
rotation too, but I keep having power system faults when jumping more
than a couple of times in a row so it is hard to validate.

Anyways, this is just a checkpoint.  Everything but continuous jumping
appears to work fine.