Explicitly command acceleration of each leg
This removes some "extra_z" hacks, and makes it possible for the lower
level controller to apply a feed forward term based on the joint
acceleration.  This doesn't end up working all that well, as the
unloaded leg isn't that well approximated by a point mass, but it is
more sound and reduces the number of special cases.

Re-work the standup logic to use this framework, as well as give it
acceleration limited tweens into the final position.