Adjust CoM location based on an estimate of the current terrain
This changes a number of things:

First, the RB offset is no longer commandable across all gaits.  There
is instead a "rest" only option to add an additional offset to that
command.

Second, there is a new "T" frame, which is an estimate of the terrain
under the robots feet.  That is estimated recursively based on feet
that are marked as being in stance and the current M frame with an
exponential filter.

Finally, all the gaits now apply a RB offset to get the CoM to be over
the center of support based on the current terrain.  This lets it walk
over the simulated ramp of size up to 0.35, (and 0.4 backwards).